+++ package.json
@@ -84,8 +84,8 @@
 		"test:blackbox:geth:ws": "yarn pre-blackbox && yarn geth:start:background && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:geth:ws --stream && yarn post-blackbox:geth",
 		"test:blackbox:infura:http": "yarn pre-blackbox && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:infura:http --stream && yarn post-blackbox",
 		"test:blackbox:infura:ws": "yarn pre-blackbox && ./scripts/verdaccio.sh startBackgroundAndPublish && lerna run test:blackbox:infura:ws --stream && yarn post-blackbox",
-		"test:stress:data": "tests/stressTests/data/start.sh",
-		"test:stress:validation": "node tests/stressTests/validation/index.js",
+		"test:stress:data": "packages/web3/test/stress/start.sh",
+		"test:stress:validation": "node packages/web3-validator/test/stress/index.js",
 		"test:stress": "yarn test:stress:data && yarn test:stress:validation",
 		"husky:install": "husky install",
 		"husky:uninstall": "husky uninstall",
