+++ packages/web3/test/stress/start.sh
@@ -8,9 +8,9 @@ export WEB3_SYSTEM_TEST_BACKEND="geth"
 yarn generate:accounts
 
 export WEB3_SYSTEM_TEST_PROVIDER=$IPC_PATH
-node ./tests/stressTests/data/index.js
+node ./packages/web3/test/stress/index.js
 
 export WEB3_SYSTEM_TEST_PROVIDER=ws://127.0.0.1:8545
-node ./tests/stressTests/data/index.js
+node ./packages/web3/test/stress/index.js
 
 ./scripts/geth_binary.sh stop
